Refactor code in order to use the Command Pattern

+++ pyscium/utils/file_util.py
@@ -1,20 +1,35 @@
-
-def open_file(filename):
-    # logger.info('open_file()')
-    file = open(filename, 'r+')
-    return file
-    # file = open(filename, 'r+', 1)
+from pathlib import Path
 
 
 def create_file(filename):
-    # logger.info('create_file')
-    file = open(filename, 'x')
-    return file
+    open(filename, 'x')
 
 
-def write_contents_to_file(string, file):
-    # logger.info("writing to file")
-    # logger.info("input: " + string)
-    file.seek(0)
-    file.write(string)
-    file.truncate()
+def open_file(filename):
+    if Path(filename).is_file():
+        file = open(filename, 'r+')
+        return file
+    else:
+        return None
+
+
+def get_file_contents(filename):
+    file = open_file(filename)
+
+    return file.readlines()
+
+
+def write_data_to_file(filename, data):
+    file = open_file(filename)
+    if file is not None:
+        file.seek(0)
+        for line in data:
+            file.write(line)
+        file.truncate()
+
+
+def close_file(filename):
+    file = open_file(filename)
+    if file is not None:
+        file.flush()
+        file.close()
